Eliminate JoltProvableTrace trait
